Refactor to ensure clearer function contracts; see commit msg
Having to use a magic string to identify profile config's filename
brought to my attention that the function design was too inconsistent to
ensure cleanliness over time. To fix this, we have the main loop
interface with functions that ask for nothing but what the user tells
it, and we do all the parsing, path building, etc., down in utility
functions.

It may seem a bit overengineered for a Bash script, but if we ever
extend this script, we'll be grateful that we did this.

Also, this is a bigger changeset than I'd normally ship in a single
commit, but there were several pieces to touch before we were assuredly
not broken.