git_profile_exec_git_validate_profile: use nameref to set config path
This is more in keeping with OOP-style validation. "Returning" a value,
even using the bash `echo` pattern, from a validator function is not
semantic and breaks expectation. Setting a variable via nameref is maybe
a weird side effect of such a function, but one that's better
precedented and easily explained by the fact that we don't know the path
value for certain _until_ we validate it.