bash-completion: few small cleanups; organize reply collections by position, not semantics
This is an attempt to set the stage for an eventual refactor where
compwords will be tracked as separate collections. The intent is to
enable easier extensibility of the script without having to make ad-hoc
changes in service of positional nuances (e.g., allowing all subcommands
in one position, allowing all but one in a subsequent position).